src/REST-interface/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
grpcio==1.80.0
typing-extensions==4.15.0
cffi
cryptography